New nvim executable path location added to setup

+++ setup/.build_win.ps1
@@ -2,6 +2,7 @@
 
 #Executables
 $nvim_exe = 'C:\Program Files\Neovim\bin\nvim.exe'
+$nvim_exe2 = 'C:\tools\neovim\nvim-win64\bin\nvim.exe'
 $posh_exe = $home +'\AppData\Local\Programs\oh-my-posh\bin\oh-my-posh.exe'
 $wezterm_exe = 'C:\Program Files\WezTerm\wezterm.exe'
 $docker_exe = 'C:\Program Files\Docker\Docker\Docker Desktop.exe'
@@ -10,15 +11,15 @@ $docker_exe = 'C:\Program Files\Docker\Docker\Docker Desktop.exe'
 $nvim_config = $home + '\AppData\Local\nvim'
 $posh_config = $home +'\Posh'
 $wezterm_config = $home + '\.wezterm.lua'
-$pwsh_config = $home + '\Documents\PowerShell'
 $docker_config = $home + '\.docker'
+$pwsh_config = split-path -parent $PROFILE
 
 #Config old
 $nvim_config_old = $nvim_config + '.old'
 $posh_config_old = $posh_config + '.old' 
 $wezterm_config_old = $wezterm_config + '.old'
 $pwsh_config_old = $pwsh_config + '.old'
-$docker_config_old = $home + '.docker' + '.old'
+$docker_config_old = $docker_config + '.old'
 
 #Fonts
 $FontFolder = $home + '\.config\fonts'
@@ -27,7 +28,8 @@ $FontList = Get-ChildItem -Path "$FontItem\*" -Include ('*.fon','*.otf','*.ttc',
 
 function Docker {
     if (Test-Path $docker_exe) {
-        if(Test-Path $docker_config){
+        if(Test-Path $docker_config) {
+            Remove-Item -Path $docker_config_old -Force -Recurse 
             Move-Item -Path $docker_config -Destination $docker_config_old
         }
         New-Item -ItemType Junction -Path $docker_config -Target (Get-Item ~\.config\.docker).FullName
@@ -37,8 +39,9 @@ function Docker {
 }
 
 function Nvim {
-    if (Test-Path $nvim_exe) {
-        if(Test-Path $nvim_config){
+    if( (Test-Path $nvim_exe) -or (Test-Path $nvim_exe2) ){
+        if(Test-Path $nvim_config) {
+            Remove-Item -Path $nvim_config_old -Force -Recurse 
             Move-Item -Path $nvim_config -Destination $nvim_config_old
         }
         New-Item -ItemType Junction -Path $nvim_config -Target (Get-Item ~\.config\nvim).FullName
@@ -49,7 +52,8 @@ function Nvim {
 
 function Posh {
     if (Test-Path $posh_exe) {
-        if(Test-Path $posh_config ){
+        if(Test-Path $posh_config ) {
+            Remove-Item -Path $posh_config_old -Force -Recurse 
             Move-Item -Path $posh_config -Destination $posh_config_old
         }
         New-Item -ItemType Junction -Path $posh_config -Target (Get-Item ~\.config\Posh).FullName
@@ -59,7 +63,8 @@ function Posh {
 }
 
 function PowerShell {
-    if(Test-Path $nvim_config){
+    if(Test-Path $pwsh_config) {
+        Remove-Item -Path $pwsh_config_old -Force -Recurse 
         Move-Item -Path $pwsh_config -Destination $pwsh_config_old
     }
     New-Item -ItemType Junction -Path $pwsh_config -Target (Get-Item ~\.config\PowerShell).FullName
@@ -67,7 +72,9 @@ function PowerShell {
 
 function WezTerm {
     if (Test-Path $wezterm_exe) {
-        if(Test-Path $wezterm_config){
+        if(Test-Path $wezterm_config) {
+
+            Remove-Item -Path  $wezterm_config_old -Force -Recurse 
             Move-Item -Path $wezterm_config -Destination $wezterm_config_old
         }
         New-Item -ItemType SymbolicLink -Path $wezterm_config -Target (Get-Item ~\.config\.wezterm.lua).FullName
@@ -77,11 +84,14 @@ function WezTerm {
 }
 function Fonts {
     foreach ($Font in $FontList) {
-        Write-Host 'Installing font -' $Font.BaseName
-        Copy-Item $Font "C:\Windows\Fonts"
-        #register font for all users
-        New-ItemProperty -Name $Font.BaseName -Path "HKLM:\Software\Microsoft\Windows NT\CurrentVersion\Fonts" -PropertyType string -Value $Font.name
-
+        $fontpath = "C:\Windows\Fonts\" + $Font
+        if([System.IO.File]::Exists($fontpath)) {
+            Write-Host 'Installing font -' $Font.BaseName
+            Copy-Item $Font "C:\Windows\Fonts"
+            New-ItemProperty -Name $Font.BaseName -Path "HKLM:\Software\Microsoft\Windows NT\CurrentVersion\Fonts" -PropertyType string -Value $Font.name
+        } else {
+            Write-Host $Font + " is already installed"
+        }
     } 
 }
 function init {
@@ -90,6 +100,7 @@ function init {
     Posh
     PowerShell
     Wezterm
+    Fonts
 }
-# init
-Fonts
+
+init
